fix: xgboost and docstring changes
