+++ bentoml/utils.py
@@ -8,6 +8,7 @@ from ._internal.environment.pip_pkg import (
     lock_pypi_versions,
     with_pip_install_options,
 )
+from .exceptions import BentoMLException
 
 __all__ = [
     "builtin_docker_image",
@@ -19,10 +20,16 @@ __all__ = [
 
 _T = t.TypeVar("_T")
 
+_REQUIRED_DOC_FIELD = ["Args", "Examples", "Returns"]
+
 
 def docstrings(docs: str):
     def decorator(func: t.Callable[..., _T]):
         func.__doc__ = inspect.cleandoc(docs)
+        if not all(i in func.__doc__ for i in _REQUIRED_DOC_FIELD):
+            raise BentoMLException(
+                f"BentoML docstring requires {', '.join(_REQUIRED_DOC_FIELD)} sections per frameworks module."
+            )
         return func
 
     return decorator
